+++ models/wrf_chem/run_scripts/RUN_IASI/IASI_O3/run_IASI_to_obsseq.ksh
@@ -12,15 +12,18 @@
   export RUN_MDART_TO_SDART=true
 #
 # SET TIME INFORMATION
-  export START_DATE=2008060406
-  export END_DATE=2008063000
+  export START_DATE=2008060106
+  export END_DATE=2008063018
   export TIME_INC=6
   export ASIM_WINDOW=3
+  export START_DATE_DATA=2008060106
 #
 # SYSTEM SPECIFIC SETTINGS
   export PROCS=8
 #  export OB_TYPE=ob_reanal
   export OB_TYPE=obs
+  export NL_APM_SCALE=1.
+  export NL_APM_SCALE_SW=.FALSE.
 #
 # INITIAL CONDITION FILES
 # Set to '1' if you want a single IC file, any other # if you want separate files (the 
@@ -49,7 +52,7 @@
   export ICBC_DIR=${ASIM_DIR}
   export HYBRID_SCRIPTS_DIR=${HYBRID_DIR}/hybrid_scripts
   export DATA_DIR=${SCRATCH_DIR}/AVE_TEST_DATA
-  export OBS_IASI_DIR=${DATA_DIR}/obs_IASI
+  export OBS_IASI_DIR=${DATA_DIR}/obs_IASI_O3
 #
 # MAKE ASSIMILATION DIRECTORY AND GO TO IT
   if [[ ! -d ${ASIM_DIR} ]]; then mkdir -p ${ASIM_DIR}; fi
@@ -167,28 +170,34 @@
         fi
 #
 # USE IASI DATA FROM PAST, PRESENT, AND NEXT DATES TO ENSURE FULL COVERAGE OF $ASIM_WINDOW 
-        rm -rf input.nml
-        rm -rf iasi_asciidata.input
-        rm -rf iasi_obs_seq.out
-        cp ${OBS_IASI_DIR}/IASIO3PROF_OBSSEQ_method2_${IASI_PAST_YYYY}${IASI_PAST_MM}${IASI_PAST_DD}.dat iasi_asciidata.input
-        cp ${DART_DIR}/observations/IASI/work/input.nml ./.
-        ${DART_DIR}/observations/IASI/work/iasi_ascii_to_obs
-        export IASI_PAST_FILE=iasi_obs_seq_${IASI_PAST_YYYY}${IASI_PAST_MM}${IASI_PAST_DD}
-        mv iasi_obs_seq.out ${IASI_PAST_FILE}
 #
+# RUN FOR PAST
+        if [[ ${L_DATE} -ne ${START_DATE_DATA} ]]; then
+           rm -rf input.nml
+           rm -rf iasi_asciidata.input
+           rm -rf iasi_obs_seq.out
+           cp ${OBS_IASI_DIR}/IASIO3PROF_OBSSEQ_method2_${IASI_PAST_YYYY}${IASI_PAST_MM}${IASI_PAST_DD}.dat iasi_asciidata.input
+           cp ${DART_DIR}/observations/IASI_O3/work/input.nml ./.
+           ${DART_DIR}/observations/IASI_O3/work/iasi_ascii_to_obs
+           export IASI_PAST_FILE=iasi_obs_seq_${IASI_PAST_YYYY}${IASI_PAST_MM}${IASI_PAST_DD}
+           mv iasi_obs_seq.out ${IASI_PAST_FILE}
+        fi
+#
+# RUN FOR PRESENT
         rm -rf input.nml
         rm -rf iasi_asciidata.input
         cp ${OBS_IASI_DIR}/IASIO3PROF_OBSSEQ_method2_${YYYY}${MM}${DD}.dat iasi_asciidata.input
-        cp ${DART_DIR}/observations/IASI/work/input.nml .
-        ${DART_DIR}/observations/IASI/work/iasi_ascii_to_obs
+        cp ${DART_DIR}/observations/IASI_O3/work/input.nml .
+        ${DART_DIR}/observations/IASI_O3/work/iasi_ascii_to_obs
         export IASI_PRES_FILE=iasi_obs_seq_${YYYY}${MM}${DD}
         mv iasi_obs_seq.out ${IASI_PRES_FILE}
 #
+# RUN FOR NEXT
         rm -rf input.nml
         rm -rf iasi_asciidata.input
         cp ${OBS_IASI_DIR}/IASIO3PROF_OBSSEQ_method2_${IASI_NEXT_YYYY}${IASI_NEXT_MM}${IASI_NEXT_DD}.dat iasi_asciidata.input
-        cp ${DART_DIR}/observations/IASI/work/input.nml .
-        ${DART_DIR}/observations/IASI/work/iasi_ascii_to_obs
+        cp ${DART_DIR}/observations/IASI_O3/work/input.nml .
+        ${DART_DIR}/observations/IASI_O3/work/iasi_ascii_to_obs
         export IASI_NEXT_FILE=iasi_obs_seq_${IASI_NEXT_YYYY}${IASI_NEXT_MM}${IASI_NEXT_DD}
         mv iasi_obs_seq.out ${IASI_NEXT_FILE}
 #
@@ -220,10 +229,16 @@
         export IASI_PAST_FILE=iasi_obs_seq_${IASI_PAST_YYYY}${IASI_PAST_MM}${IASI_PAST_DD}
         export IASI_PRES_FILE=iasi_obs_seq_${YYYY}${MM}${DD}
         export IASI_NEXT_FILE=iasi_obs_seq_${IASI_NEXT_YYYY}${IASI_NEXT_MM}${IASI_NEXT_DD}
-        if ${RUN_PREP_ASCII_TO_DART}; then cp ${ASIM_DIR}/prep_ascii_to_dart/${YYYY}${MM}/obs_seq${D_DATE} ./.; fi
-        if ${RUN_IASI_ASCII_TO_DART}; then cp ${ASIM_DIR}/IASI_ascii_to_dart/${YYYY}${MM}/${IASI_PAST_FILE} ./.; fi
-        if ${RUN_IASI_ASCII_TO_DART}; then cp ${ASIM_DIR}/IASI_ascii_to_dart/${YYYY}${MM}/${IASI_PRES_FILE} ./.; fi
-        if ${RUN_IASI_ASCII_TO_DART}; then cp ${ASIM_DIR}/IASI_ascii_to_dart/${YYYY}${MM}/${IASI_NEXT_FILE} ./.; fi
+        if ${RUN_PREP_ASCII_TO_DART}; then 
+           cp ${ASIM_DIR}/prep_ascii_to_dart/${YYYY}${MM}/obs_seq${D_DATE} ./.
+        fi
+        if ${RUN_IASI_ASCII_TO_DART}; then
+           if [[ ${L_DATE} -ne ${START_DATE_DATA} ]]; then 
+              cp ${ASIM_DIR}/IASI_ascii_to_dart/${YYYY}${MM}/${IASI_PAST_FILE} ./.
+           fi
+           cp ${ASIM_DIR}/IASI_ascii_to_dart/${YYYY}${MM}/${IASI_PRES_FILE} ./.
+           cp ${ASIM_DIR}/IASI_ascii_to_dart/${YYYY}${MM}/${IASI_NEXT_FILE} ./.
+        fi
 #
 # CALCULATE GREGORIAN TIMES FOR START AND END OF ASSIMILAtION WINDOW
         export ASIM_MIN_DATE=$($BUILD_DIR/da_advance_time.exe $L_DATE -$ASIM_WINDOW 2>/dev/null)
@@ -239,24 +254,37 @@
         if ${RUN_PREP_ASCII_TO_DART}; then 
            export NL_NUM_INPUT_FILES=4
            export NL_FILENAME_SEQ="'obs_seq${D_DATE}','${IASI_PAST_FILE}','${IASI_PRES_FILE}','${IASI_NEXT_FILE}'"
-        else
+        elif [[  ${L_DATE} -ne ${START_DATE_DATA} ]]; then
            export NL_NUM_INPUT_FILES=3
            export NL_FILENAME_SEQ="'${IASI_PAST_FILE}','${IASI_PRES_FILE}','${IASI_NEXT_FILE}'"
+        else
+           export NL_NUM_INPUT_FILES=2
+           export NL_FILENAME_SEQ="'${IASI_PRES_FILE}','${IASI_NEXT_FILE}'"
         fi
         export NL_FILENAME_OUT="'obs_seq_${L_DATE}.out'"
         export NL_FIRST_OBS_DAYS=${ASIM_MIN_DAY_GREG}
         export NL_FIRST_OBS_SECONDS=${ASIM_MIN_SEC_GREG}
         export NL_LAST_OBS_DAYS=${ASIM_MAX_DAY_GREG}
         export NL_LAST_OBS_SECONDS=${ASIM_MAX_SEC_GREG}
-        export NL_SYNONYMOUS_COPY_LIST="'NCEP BUFR observation','observation'"
-        export NL_SYNONYMOUS_QC_LIST="'NCEP QC index','Data QC'"
+        export NL_SYNONYMOUS_COPY_LIST="'IASI O3 observation'"
+        export NL_SYNONYMOUS_QC_LIST="'IASI O3 QC index'"
         ${HYBRID_SCRIPTS_DIR}/da_create_dart_input_nml.ksh       
+#
+# Make obs_def_apm_nml for apm_scale to adjust observation error variance
+        rm -rf obs_def_apm.nml
+        cat <<EOF > obs_def_apm.nml
+&obs_def_apm_nml
+apm_scale=${NL_APM_SCALE}
+apm_scale_sw=${NL_APM_SCALE_SW}
+/
+EOF
+#
         ./obs_sequence_tool
-        if [[ ! -d ${DATA_DIR}/obs_IASI/${L_DATE} ]]; then mkdir -p ${DATA_DIR}/obs_IASI/${L_DATE}; fi
+        if [[ ! -d ${DATA_DIR}/obs_IASI_O3/${L_DATE} ]]; then mkdir -p ${DATA_DIR}/obs_IASI_O3/${L_DATE}; fi
         if [[ -e obs_seq_${L_DATE}.out ]]; then
-           cp obs_seq_${L_DATE}.out ${DATA_DIR}/obs_IASI/${L_DATE}/.
+           cp obs_seq_${L_DATE}.out ${DATA_DIR}/obs_IASI_O3/${L_DATE}/.
         else
-           touch ${DATA_DIR}/obs_IASI/${L_DATE}/NO_OBS_SEQ.OUT_DATA
+           touch ${DATA_DIR}/obs_IASI_O3/${L_DATE}/NO_OBS_SEQ.OUT_DATA
         fi
         cd ${ASIM_DIR}
      fi
